write ReadSector, and boost performance
I realized that, as long as the write operation for AddSector syncs
before the COW operation is performed on the WAL, and as long as the
AddSector operation is not overwriting data which was removed from an
uncommitted RemoveSector call, that there is no threat to consistency.

This means that large amounts of sector data actually never need to be
written to the WAL, which means the most significant scalability
limitation has been removed from the contract manager.

From my best understanding, the contract manager should now be set to
scale to contracts as large as 100 GiB and total volumes as large as 1
PiB.

The bottleneck for larger contracts is the renew operation. Further
optimization around turning the renew operation into a long-running
COW-based operation instead of a series of indiviual sector
modifications should get scalability improvements by as much as 1000x,
enabling 100 TiB contracts (but still bottlenecked a 1 PiB total drive
size). The performance key there is to write all of the commitments to
the WAL at once, before putting them to disk. Then, the operation that
applies the operations from the WAL to the disk can take several
minutes, but the storage obligation can return right away.

There will still be a limit to how much can be renewed per second, but
this is a total number of sectors. The current bottleneck is that a
renew operation must take place inside of a few seconds. As long as the
commitment happens inside of a few seconds, it should be okay for the
application to actually take a few minutes.

That is a DoS vector though, as someone may be inclined to renew a large
contract a large number of times with the intention of freezing the
host.